[Task]: Apply native parameter and return type hints according to php docs

+++ tests/Ecommerce/CartTaxManagementTest.php
@@ -43,7 +43,7 @@ class CartTaxManagementTest extends EcommerceTestCase
     private function buildTaxClass(array $taxes = [], $combinationType = TaxEntry::CALCULATION_MODE_COMBINE): OnlineShopTaxClass
     {
         $taxClass = new OnlineShopTaxClass();
-        $taxClass->setId(md5(serialize($taxes)));
+        $taxClass->setId((int)md5(serialize($taxes)));
 
         $taxEntries = new Fieldcollection();
         foreach ($taxes as $name => $tax) {
